feat: junta anamnesis
